+++ packages/core/agent/src/launcher/launchctl-daemon.ts
@@ -5,23 +5,30 @@
 import { existsSync, mkdirSync } from 'node:fs';
 import { readdir } from 'node:fs/promises';
 import path from 'node:path';
+import { platform } from 'node:os';
 
 import { waitForCondition } from '@dxos/async';
 import { log } from '@dxos/log';
 
-import { LaunchctlRunner } from './launchctl-runner';
+import { LaunchctlRunner, Runner } from './runner';
 import { Daemon, ProcessInfo, StartOptions, StopOptions } from '../daemon';
 import { CHECK_INTERVAL, DAEMON_STOP_TIMEOUT } from '../defs';
 import { AgentWaitTimeoutError } from '../errors';
 import { lockFilePath, removeLockFile, removeSocketFile, waitForAgentToStart } from '../util';
 
 /**
- * Manager of system daemon processes using launchctl on macOS.
+ * Manager of system daemon processes using system service manager.
  */
-export class LaunchctlDaemon implements Daemon {
-  private readonly _launchctlRunner = new LaunchctlRunner();
+export class SystemDaemon implements Daemon {
+  private readonly _runner: Runner;
 
   constructor(private readonly _rootDir: string) {
+    if (platform() === 'darwin') {
+      this._runner = new LaunchctlRunner();
+    } else {
+      throw new Error(`System daemon not implemented for ${platform()} yet.`);
+    }
+
     const dir = path.join(this._rootDir, 'profile');
     if (!existsSync(dir)) {
       mkdirSync(dir, { recursive: true });
@@ -40,7 +47,7 @@ export class LaunchctlDaemon implements Daemon {
     const { isLocked } = await import('@dxos/client-services');
     // TODO(egorgripasov): Better compatibility with Phoenix daemon.
     // Support of existing locks?
-    return (await isLocked(lockFilePath(profile))) || this._launchctlRunner.isRunning(`org.dxos.agent.${profile}`);
+    return (await isLocked(lockFilePath(profile))) || this._runner.isRunning(profile);
   }
 
   async list(): Promise<ProcessInfo[]> {
@@ -71,16 +78,7 @@ export class LaunchctlDaemon implements Daemon {
       const logFile = path.join(logDir, 'daemon.log');
       const errFile = path.join(logDir, 'err.log');
 
-      // TODO(egorgripasov): Pass options to the agent start command.
-      const plist = plistTemplate
-        .replace('{{PROFILE}}', profile)
-        // TODO(egorgripasov): Make it work for standalone installation.
-        .replace('{{DX_PATH}}', process.argv[1])
-        .replace('{{NODE_PATH}}', path.dirname(process.execPath))
-        .replace('{{ERROR_LOG}}', errFile)
-        .replace('{{OUT_LOG}}', logFile);
-
-      await this._launchctlRunner.load(`org.dxos.agent.${profile}`, plist);
+      await this._runner.start({ profile, errFile, logFile });
 
       try {
         await waitForAgentToStart(profile);
@@ -98,8 +96,7 @@ export class LaunchctlDaemon implements Daemon {
     if (await this.isRunning(profile)) {
       log('stopping...', { profile });
 
-      // Use your LaunchctlRunner to unload the .plist file.
-      await this._launchctlRunner.unload(`org.dxos.agent.${profile}`, force);
+      await this._runner.stop(profile, force);
 
       await waitForCondition({
         condition: async () => !(await this.isRunning(profile)),
@@ -117,41 +114,3 @@ export class LaunchctlDaemon implements Daemon {
     return this.start(profile, options);
   }
 }
-
-// TODO(egorgripasov): Read from template file.
-const plistTemplate = `<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
-<plist version="1.0">
-    <dict>
-    <key>Label</key>
-    <string>org.dxos.agent.{{PROFILE}}</string>
-    <key>KeepAlive</key>
-    <true/>
-    <key>ProgramArguments</key>
-    <array>
-      <string>{{DX_PATH}}</string>
-      <string>agent</string>
-      <string>start</string>
-      <string>--foreground</string>
-    </array>
-    <key>RunAtLoad</key>
-    <true/>
-    <key>KeepAlive</key>
-    <dict>
-      <key>Crashed</key>
-      <true/>
-    </dict>
-    <key>EnvironmentVariables</key>
-    <dict>
-      <key>PATH</key>
-      <string>{{NODE_PATH}}</string>
-      <key>LOG_FILTER</key>
-      <string>info</string>
-    </dict>
-    <key>StandardErrorPath</key>
-    <string>{{ERROR_LOG}}</string>
-    <key>StandardOutPath</key>
-    <string>{{OUT_LOG}}</string>
-  </dict>
-</plist>
-`;
